Fix: Travel의 grade 타입 변경

+++ SeSAC7Week3_T1/Model/Travel.swift
@@ -7,7 +7,7 @@ struct Travel {
     let title: String
     let description: String
     let travel_image: String
-    let grade: Int
+    let grade: Double
     let save: Int
     let like: Bool
     let ad: Bool
